remove remainder of middlewared.schema
