+++ src/middlewared/middlewared/plugins/network_/dns.py
@@ -13,8 +13,7 @@ from middlewared.api.current import DNSQueryItem
 from middlewared.service import Service, filterable_api_method, private
 from middlewared.utils import filter_list, MIDDLEWARE_RUN_DIR
 from middlewared.plugins.interface.netif import netif
-from middlewared.schema import ValidationErrors
-from middlewared.service_exception import CallError
+from middlewared.service_exception import CallError, ValidationErrors
 
 
 class DNSNsUpdateOpA(BaseModel):
