+++ src/middlewared/middlewared/plugins/vm/devices/device.py
@@ -1,7 +1,7 @@
 from abc import ABC
 
 from middlewared.api.base.handler.accept import validate_model
-from middlewared.schema import ValidationErrors
+from middlewared.service_exception import ValidationErrors
 
 
 class Device(ABC):
