--- src/middlewared/middlewared/schema/__init__.py
@@ -1,7 +1,0 @@
-from middlewared.service_exception import ValidationErrors  # noqa
-
-from .attribute import Attribute  # noqa
-from .exceptions import Error  # noqa
-from .plugin_schema import Schemas  # noqa
-from .resolvers import resolve_methods  # noqa
-from .utils import NOT_PROVIDED, REDACTED_VALUE  # noqa
