--- src/middlewared/middlewared/schema/utils.py
@@ -1,2 +1,0 @@
-NOT_PROVIDED = object()
-REDACTED_VALUE = "********"
